requirements/base.txt
Django==1.5.1
psycopg2==2.5
path.py==4.1
south==0.8.1
django-pipeline==1.3.11
django-extensions==1.1.1
https://bitbucket.org/ubernostrum/django-registration/get/b3c41b3.tar.gz#egg=django_registration
# Install django-oscar (latest version, compatible with Django 1.5)
git+http://github.com/tangentlabs/django-oscar.git@issue/472/django1.5#egg=django-oscar
django-oscar-paypal==0.4.1
django-model-utils==1.4.0
django-storages==1.1.8
boto==2.9.5
sorl-thumbnail==11.12
django-tastypie==0.9.15
https://github.com/adamcharnock/django-tz-detect/archive/v0.1.4.tar.gz#egg=django-tz-detect
-e git+https://github.com/adamcharnock/manifesto.git#egg=manifesto
fabric
django-celery-with-redis==3.0